+++ src/main.cpp
@@ -1,4 +1,4 @@
-#include "headers/Logs.hpp"
+#include "./Debugger.hpp"
 
 #include "../incl/Avalanche.hpp"